ofxSynthDelayLine: check for size > than 2 secs + use floats instead of doubles. Closes #457

+++ addons/ofxSynth/src/ofxSynthDelayLine.h
@@ -13,12 +13,12 @@ class ofxSynthDelayline : public ofSoundEffect {
 		void setMix(float _mix);
 		
 	private:
-		double frequency;
+		float frequency;
 		int phase;
-		double startphase;
-		double endphase;
-		double output;
-		double memory[88200];
+		float startphase;
+		float endphase;
+		float output;
+		float memory[88200];
 		int size;
 		float feedback;
 		float mix;
